Reorganized files into package structure
