+++ requirements-dev.txt
@@ -1,0 +1,4 @@
+build
+setuptools
+setuptools_scm
+wheel
